PhoneSale-main/backend/requirements.txt
Django==5.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
lazr.restfulclient==0.14.6
pillow==12.1.1
PyJWT==2.7.0
